pyproject.toml
[tool.poetry]
name = "article-recs"
version = "0.1.0"
description = ""
authors = ["dimafeng <dimafeng@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
SQLAlchemy = "^1.4.44"
psycopg2-binary = "^2.9.5"
newspaper3k = "^0.2.8"
python-telegram-bot = "^13.14"
schedule = "^1.1.0"
tenacity = "^8.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
